DFPL-1475: Duplicate notifications being sent to CTSC (#4458)
* DFPL-1475: Duplicate notifications being sent to CTSC

* remove unused import

* update main approvedraftorders tests

* update tests with timeout for new async call

* checkstyle

* add timeouts

* add check for events in case

* checkstyle

* additional async testing changes

* additional async testing changes +

* checkstyle

* add constant and checkstyle changes

* checkstyle + additional checks for no interactions

* remove whens from no interaction tests

* remove unused import

* Potential idea to restructure assertions

* add ordering to prevent no interactions failing

* checkstyle

* add ordering to no interaction tests

* add ordering for failing no interaction test

* add more ordering to test

* suppressions

* no more interactions

* add verification for start events

* remove unused import and add timeouts

* add more timeouts

* re-trigger integration tests

* re-trigger integration tests 2

* add additional logging for async tests

* add additional logging for async tests 2

* add additional logging for async tests 2

* Empty-Commit

* remove old api tests (accident)

* Empty-Commit

* generate new instance of caseDetails to prevent concurrent modification exception

* additional timeouts

* Empty-Commit

* Empty-Commit

* check smoketests

* smoketests add back silent

* remove logging from tests and add more timeouts

* remove checkUntil

* Empty-Commit

* Empty-Commit

* Empty-Commit

* add captor for shouldNotTrigger events

* Empty-Commit

* Empty-Commit

* add comment to build gradle

---------
