+++ service/build.gradle
@@ -1,4 +1,4 @@
- buildscript {
+buildscript {
   repositories {
     mavenCentral()
   }
@@ -54,51 +54,59 @@ sourceSets {
     }
   }
 
-    smokeTest {
-        java {
-            compileClasspath += main.output
-            runtimeClasspath += main.output
-            srcDir file('src/smokeTest/java')
-        }
+  smokeTest {
+    java {
+      compileClasspath += main.output
+      runtimeClasspath += main.output
+      srcDir file('src/smokeTest/java')
     }
+  }
 
-    contractTest {
-        java {
-            compileClasspath += main.output
-            runtimeClasspath += main.output
-            srcDir file('src/contractTest/java')
-        }
+  contractTest {
+    java {
+      compileClasspath += main.output
+      runtimeClasspath += main.output
+      srcDir file('src/contractTest/java')
     }
+  }
 }
 
 idea {
- module {
-   testSourceDirs += file('src/integrationTest/java')
-   testSourceDirs += file('src/functionalTest/java')
- }
+  module {
+    testSourceDirs += file('src/integrationTest/java')
+    testSourceDirs += file('src/functionalTest/java')
+  }
 }
 
 configurations {
- functionalTestImplementation.extendsFrom testImplementation
- functionalTestRuntime.extendsFrom testRuntime
- integrationTestImplementation.extendsFrom testImplementation
- integrationTestRuntime.extendsFrom testRuntime
- cftlibTestImplementation.extendsFrom testImplementation
- cftlibTestRuntime.extendsFrom testRuntime
+  functionalTestImplementation.extendsFrom testImplementation
+  functionalTestRuntime.extendsFrom testRuntime
+  integrationTestImplementation.extendsFrom testImplementation
+  integrationTestRuntime.extendsFrom testRuntime
+  cftlibTestImplementation.extendsFrom testImplementation
+  cftlibTestRuntime.extendsFrom testRuntime
 }
 
 configurations.integrationTestImplementation {
- exclude group: 'org.springframework.cloud', module: 'spring-cloud-starter-netflix-hystrix'
+  exclude group: 'org.springframework.cloud', module: 'spring-cloud-starter-netflix-hystrix'
 }
 
 tasks.withType(JavaCompile) {
   options.compilerArgs << "-Xlint:unchecked" << "-Werror"
 }
 
+//Comment this section in to allow for additional logging of unit tests for any tests with type 'Test'
+/*tasks.withType(Test) {
+  useJUnitPlatform()
+
+  testLogging {
+    exceptionFormat = 'full'
+  }
+}*/
+
 task integration(type: Test, description: 'Runs the integration tests.', group: 'Verification') {
   testClassesDirs = sourceSets.integrationTest.output.classesDirs
   classpath = sourceSets.integrationTest.runtimeClasspath
-
   // set your environment variables here
   // environment("APPINSIGHTS_INSTRUMENTATIONKEY", "test-key")
 }
@@ -123,25 +131,25 @@ task runApiTests(type: Test, description: 'Runs api tests.', group: 'Verificatio
 }
 
 project.ext {
-    pacticipant = 'fpl_ccdConfiguration'
-    pacticipantVersion = getCheckedOutGitCommitHash()
+  pacticipant = 'fpl_ccdConfiguration'
+  pacticipantVersion = getCheckedOutGitCommitHash()
 }
 
 task runContractTests(type: Test, description: 'Runs contract tests.', group: 'Verification') {
-    onlyIf {
-        return System.env.SKIP_CONTRACT_TESTS == null || System.env.SKIP_CONTRACT_TESTS == 'false'
-    }
-    useJUnitPlatform()
-    testClassesDirs = sourceSets.contractTest.output.classesDirs
-    classpath = sourceSets.contractTest.runtimeClasspath
-    systemProperty 'pact.rootDir', "pacts"
-    testLogging.showStandardStreams = true
+  onlyIf {
+    return System.env.SKIP_CONTRACT_TESTS == null || System.env.SKIP_CONTRACT_TESTS == 'false'
+  }
+  useJUnitPlatform()
+  testClassesDirs = sourceSets.contractTest.output.classesDirs
+  classpath = sourceSets.contractTest.runtimeClasspath
+  systemProperty 'pact.rootDir', "pacts"
+  testLogging.showStandardStreams = true
 }
 
 task runAndPublishConsumerPactTests(type: Test, description: 'Publish pact test to broker.', group: 'Verification') {
-    logger.lifecycle("Runs pact consumer Tests")
-    testClassesDirs = sourceSets.contractTest.output.classesDirs
-    classpath = sourceSets.contractTest.runtimeClasspath
+  logger.lifecycle("Runs pact consumer Tests")
+  testClassesDirs = sourceSets.contractTest.output.classesDirs
+  classpath = sourceSets.contractTest.runtimeClasspath
 }
 
 runAndPublishConsumerPactTests.dependsOn runContractTests
@@ -149,18 +157,18 @@ runAndPublishConsumerPactTests.dependsOn runContractTests
 runAndPublishConsumerPactTests.finalizedBy pactPublish
 
 def getCheckedOutGitCommitHash() {
-    'git rev-parse --verify --short HEAD'.execute().text.trim()
+  'git rev-parse --verify --short HEAD'.execute().text.trim()
 }
 
 pact {
-    broker {
-        pactBrokerUrl = System.getenv("PACT_BROKER_FULL_URL") ?: 'http://localhost:80'
-    }
-    publish {
-        pactDirectory = 'service/pacts'
-        tags = [System.getenv("PACT_BRANCH_NAME") ?:'Dev']
-        version = project.pacticipantVersion
-    }
+  broker {
+    pactBrokerUrl = System.getenv("PACT_BROKER_FULL_URL") ?: 'http://localhost:80'
+  }
+  publish {
+    pactDirectory = 'service/pacts'
+    tags = [System.getenv("PACT_BRANCH_NAME") ?: 'Dev']
+    version = project.pacticipantVersion
+  }
 }
 
 
@@ -188,7 +196,7 @@ task functional(description: 'Runs the functional tests.') {
 }
 
 task contract(description: 'Runs the contract tests.') {
-    dependsOn(inStrictOrder(runContractTests))
+  dependsOn(inStrictOrder(runContractTests))
 }
 
 task runCrossBrowserTests(type: Exec, description: 'Runs ExpertUI cross brower tests.') {
@@ -202,7 +210,7 @@ task crossbrowser(description: 'Runs crossbrowser tests.') {
   dependsOn(inStrictOrder(lintUserInterfaceTests, runCrossBrowserTests))
 }
 
-task fortifyScan(type: JavaExec)  {
+task fortifyScan(type: JavaExec) {
   mainClass = "uk.gov.hmcts.fortifyclient.FortifyClientMainApp"
   classpath += sourceSets.test.runtimeClasspath
   jvmArgs = ['--add-opens=java.base/java.lang.reflect=ALL-UNNAMED']
@@ -343,7 +351,7 @@ dependencies {
 
   implementation group: 'com.fasterxml.jackson.datatype', name: 'jackson-datatype-joda', version: '2.12.4'
 
-  implementation group: 'net.minidev', name:'json-smart', version: '2.4.7'
+  implementation group: 'net.minidev', name: 'json-smart', version: '2.4.7'
 
   implementation group: 'org.springframework.boot', name: 'spring-boot-starter-mail'
 
@@ -378,7 +386,6 @@ dependencies {
   functionalTestAnnotationProcessor group: 'org.projectlombok', name: 'lombok', version: versions.lombok
   functionalTestImplementation sourceSets.main.runtimeClasspath
   functionalTestImplementation sourceSets.test.runtimeClasspath
-
 
 
   integrationTestImplementation sourceSets.main.runtimeClasspath
@@ -443,7 +450,7 @@ pitest {
     'uk.gov.hmcts.reform.fpl.controllers.*',
     'uk.gov.hmcts.reform.fpl.testingsupport.*'
   ]
-  jvmArgs = ['-Xms1G','-Xmx3G']
+  jvmArgs = ['-Xms1G', '-Xmx3G']
   threads = 10
   enableDefaultIncrementalAnalysis = true
   outputFormats = ['XML', 'HTML']
@@ -494,7 +501,7 @@ bootWithCCD {
 
   doFirst() {
     project.file('../build/cftlib/.aat-env').readLines().each() {
-      if (it.contains("=")){
+      if (it.contains("=")) {
         def index = it.indexOf("=")
         def key = it.substring(0, index)
         def value = it.substring(index + 1)
