+++ service/src/integrationTest/java/uk/gov/hmcts/reform/fpl/controllers/MessageJudgeControllerSubmittedTest.java
@@ -21,6 +21,7 @@ import java.util.Map;
 
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.ArgumentMatchers.eq;
+import static org.mockito.Mockito.timeout;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 import static uk.gov.hmcts.reform.fpl.Constants.LOCAL_AUTHORITY_1_CODE;
@@ -108,7 +109,7 @@ class MessageJudgeControllerSubmittedTest extends AbstractCallbackTest {
 
         verify(notificationClient).sendEmail(
             JUDICIAL_MESSAGE_ADDED_TEMPLATE, JUDICIAL_MESSAGE_RECIPIENT, expectedData, notificationReference(CASE_ID));
-        verify(concurrencyHelper).submitEvent(any(),
+        verify(concurrencyHelper, timeout(10000)).submitEvent(any(),
             eq(CASE_ID),
             eq(caseSummary()));
     }
