+++ service/src/integrationTest/java/uk/gov/hmcts/reform/fpl/controllers/RespondentControllerSubmittedTest.java
@@ -2,7 +2,10 @@ package uk.gov.hmcts.reform.fpl.controllers;
 
 import org.apache.commons.lang3.RandomStringUtils;
 import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.MethodOrderer;
+import org.junit.jupiter.api.Order;
 import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.TestMethodOrder;
 import org.mockito.AdditionalAnswers;
 import org.mockito.ArgumentCaptor;
 import org.mockito.Captor;
@@ -31,6 +34,7 @@ import uk.gov.hmcts.reform.rd.client.OrganisationApi;
 import uk.gov.service.notify.NotificationClient;
 import uk.gov.service.notify.NotificationClientException;
 
+import java.time.Duration;
 import java.time.LocalDate;
 import java.util.List;
 import java.util.Map;
@@ -42,9 +46,11 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.ArgumentMatchers.anyMap;
 import static org.mockito.ArgumentMatchers.eq;
+import static org.mockito.Mockito.timeout;
 import static org.mockito.Mockito.times;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.verifyNoInteractions;
+import static org.mockito.Mockito.verifyNoMoreInteractions;
 import static org.mockito.Mockito.when;
 import static uk.gov.hmcts.reform.ccd.model.ChangeOrganisationApprovalStatus.APPROVED;
 import static uk.gov.hmcts.reform.ccd.model.Organisation.organisation;
@@ -59,6 +65,7 @@ import static uk.gov.hmcts.reform.fpl.enums.State.SUBMITTED;
 import static uk.gov.hmcts.reform.fpl.enums.YesNo.YES;
 import static uk.gov.hmcts.reform.fpl.handlers.NotificationEventHandlerTestData.COURT_NAME;
 import static uk.gov.hmcts.reform.fpl.service.ccd.CoreCaseDataService.UPDATE_CASE_EVENT;
+import static uk.gov.hmcts.reform.fpl.utils.AssertionHelper.checkThat;
 import static uk.gov.hmcts.reform.fpl.utils.AssertionHelper.checkUntil;
 import static uk.gov.hmcts.reform.fpl.utils.ElementUtils.element;
 import static uk.gov.hmcts.reform.fpl.utils.ElementUtils.wrapElements;
@@ -66,6 +73,7 @@ import static uk.gov.hmcts.reform.fpl.utils.TestDataHelper.caseRoleDynamicList;
 
 @WebMvcTest(RespondentController.class)
 @OverrideAutoConfiguration(enabled = true)
+@TestMethodOrder(MethodOrderer.OrderAnnotation.class)
 class RespondentControllerSubmittedTest extends AbstractCallbackTest {
 
     private static final State NON_RESTRICTED_STATE = SUBMITTED;
@@ -75,7 +83,9 @@ class RespondentControllerSubmittedTest extends AbstractCallbackTest {
     private static final String SOLICITOR_ORG_ID = "Organisation ID";
     private static final String SOLICITOR_EMAIL = "solicitor@email.com";
     private static final String CASE_ID = "1234567890123456";
+    private static final long CASE_ID_LONG = 1234567890123456L;
     private static final String NOTIFICATION_REFERENCE = "localhost/" + CASE_ID;
+    private static final long ASYNC_METHOD_CALL_TIMEOUT = 10000;
 
     private final Organisation organisation1 = organisation("ORG_1");
     private final Organisation organisation2 = organisation("ORG_2");
@@ -113,6 +123,9 @@ class RespondentControllerSubmittedTest extends AbstractCallbackTest {
 
     @Captor
     private ArgumentCaptor<Map<String, Object>> caseCaptor;
+
+    @Captor
+    private ArgumentCaptor<StartEventResponse> startEventCaptor;
 
     RespondentControllerSubmittedTest() {
         super("enter-respondents");
@@ -195,6 +208,10 @@ class RespondentControllerSubmittedTest extends AbstractCallbackTest {
             anyMap(),
             eq(NOTIFICATION_REFERENCE)
         ));
+
+        verify(concurrencyHelper, timeout(ASYNC_METHOD_CALL_TIMEOUT)).startEvent(eq(CASE_ID_LONG), any());
+        verify(concurrencyHelper, timeout(ASYNC_METHOD_CALL_TIMEOUT)).submitEvent(any(), eq(CASE_ID_LONG), anyMap());
+        verifyNoMoreInteractions(concurrencyHelper);
     }
 
     @Test
@@ -320,6 +337,12 @@ class RespondentControllerSubmittedTest extends AbstractCallbackTest {
         verify(notificationClient).sendEmail(
             LEGAL_COUNSELLOR_REMOVED_EMAIL_TEMPLATE, legalCounsellorEmail, notifyData, "localhost/" + CASE_ID
         );
+        verify(concurrencyHelper, timeout(ASYNC_METHOD_CALL_TIMEOUT).times(3))
+            .startEvent(eq(CASE_ID_LONG), any());
+        verify(concurrencyHelper, timeout(ASYNC_METHOD_CALL_TIMEOUT).times(3))
+            .submitEvent(any(), eq(CASE_ID_LONG), anyMap());
+
+        verifyNoMoreInteractions(concurrencyHelper);
     }
 
     @Test
@@ -364,33 +387,38 @@ class RespondentControllerSubmittedTest extends AbstractCallbackTest {
             .map(changeRequest -> Map.of("changeOrganisationRequestField", changeRequest))
             .collect(toList());
 
+        final List<String> expectedCaseEvents = List.of("internal-update-case-summary",
+            "internal-change-UPDATE_CASE", "updateRepresentation");
+
+        verify(concurrencyHelper, timeout(ASYNC_METHOD_CALL_TIMEOUT).times(5))
+            .startEvent(eq(caseData.getId()), any());
+
         // setup change field, submit to aac (org1), setup change field, submit to aac (org2), update case-summary
-        verify(concurrencyHelper, times(5)).submitEvent(any(),
-            eq(caseData.getId()), caseCaptor.capture()
-        );
+        checkThat(() -> {
+            verify(concurrencyHelper, times(5)).submitEvent(startEventCaptor.capture(),
+                eq(caseData.getId()), caseCaptor.capture()
+            );
+        }, Duration.ofSeconds(10));
+
+        List<String> temp = startEventCaptor.getAllValues().stream()
+            .map(StartEventResponse::getEventId).collect(toList());
+
+        assertThat(temp).asList().containsAll(expectedCaseEvents);
 
         assertThat(caseCaptor.getAllValues())
             .asList()
-            .containsOnlyOnceElementsOf(update);
+            .containsAll(update);
+
+        verifyNoMoreInteractions(concurrencyHelper);
     }
 
     @Test
+    @Order(1)
     void shouldNotUpdateRepresentativesAccessWhenCaseNotSubmitted() {
         final CaseData caseData = nocCaseDataBefore.toBuilder()
             .state(OPEN)
             .respondents1(updatedRespondents)
             .build();
-
-        when(concurrencyHelper.startEvent(any(), any()))
-            .thenReturn(StartEventResponse.builder()
-                .caseDetails(asCaseDetails(caseData))
-                .eventId(UPDATE_CASE_EVENT)
-                .build());
-        when(concurrencyHelper.startEvent(any(), eq("internal-update-case-summary")))
-            .thenReturn(StartEventResponse.builder()
-                .caseDetails(asCaseDetails(caseData))
-                .eventId("internal-update-case-summary")
-                .build());
 
         postSubmittedEvent(toCallBackRequest(caseData, nocCaseDataBefore));
