+++ service/src/main/java/uk/gov/hmcts/reform/fpl/service/ccd/CoreCaseDataService.java
@@ -37,6 +37,11 @@ public class CoreCaseDataService {
     private final SystemUserService systemUserService;
     private final CCDConcurrencyHelper concurrencyHelper;
 
+    public CaseDetails performPostSubmitCallbackWithoutChange(Long caseId, String eventName) {
+        return performPostSubmitCallback(caseId, eventName,
+            (caseDetaills) -> Map.of(), true);
+    }
+
     public CaseDetails performPostSubmitCallback(Long caseId,
                                                  String eventName,
                                                  Function<CaseDetails, Map<String, Object>> changeFunction) {
